requirements.txt
discord.py
requests
termcolor